Move the quickcheck extras into a private module called smallcheck
