+++ src/lib.rs
@@ -2,5 +2,6 @@
 #[macro_use]
 extern crate quickcheck;
 
+mod smallcheck;
 pub mod traits;
 pub mod types;
